Added class files
