+++ GameState.py
@@ -1,0 +1,3 @@
+from enum import Enum
+class GameState(Enum):
+    QUIT = -1
